feat: add settings input to set api endpoint
